Changed definition of atom to specifically be nat
This seems to move feval closer to fully computing the result,
but it seems that some atoms do not actually have their values
defined.  I was able to get the lookup function to fully evaluate
by defining the values for the atoms that it compared, like so:

let Pair := 1 in let A := 2 in let B:=3 in lookup Pair
  ((A,(Object,nil,nil)) ::(B,(Object,nil,nil))
  ::(Pair,(Object,pair_flds,pair_mths)) :: nil)

However, feval still does not fully evaluate, which may mean that
some atom is still not defined.  I tried defining all the atoms
that I could find that it might compare, but it did not work.